\name{correlation}
\alias{correlation}
\title{Correlation}
\usage{
  correlation(x, y)
}
\arguments{
  \item{x}{a numeric vector}

  \item{y}{another numeric vector of length equal to 'x'}
}
\value{
  a correlation coefficient.
}
\description{
  Calcualte a correlation coefficient between two vectors
}

