\name{linear}
\alias{linear}
\title{Linear Regression}
\usage{
  linear(x, y)
}
\arguments{
  \item{x}{numeric vector on x axis.}

  \item{y}{numeric vector on y axis.}
}
\value{
  intercept and slope
}
\description{
  Perform a linear regression of two variables (i.e.
  calculate optimal slope).
}

