\name{quantiles}
\alias{quantiles}
\title{Quantiles}
\usage{
  quantiles(x, probs = c(0, 0.25, 0.5, 0.75, 1))
}
\arguments{
  \item{x}{a numeric vector.}

  \item{probs}{vector with quantie probabilities. All
  values must be between 0 and 1.}
}
\value{
  probabilities and corresponding quantiles.
}
\description{
  Calculate quantiles for a numeric vector
}

